requirements.txt
                  
langchain==0.0.310
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.27.2
openai==0.27.8
altair==5.1.2
tiktoken==0.5.2
